+++ composeApp/src/androidMain/kotlin/com/yong/taximeter/common/util/PreferenceUtil.android.kt
@@ -8,11 +8,16 @@ import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
-
-private val Context.dataStore by preferencesDataStore("app_preferences")
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
+import kotlin.getValue
 
 @Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-actual class PreferenceManager(private val context: Context) {
+actual object PreferenceUtil: KoinComponent {
+    // Android Context
+    private val context: Context by inject()
+    private val Context.dataStore by preferencesDataStore("app_preferences")
+
     actual suspend fun getString(key: String, defaultValue: String): String {
         return context.dataStore.data.map { preferences ->
             preferences[stringPreferencesKey(key)] ?: defaultValue
