+++ composeApp/src/commonMain/kotlin/com/yong/taximeter/common/def/PreferenceDef.kt
@@ -1,0 +1,8 @@
+package com.yong.taximeter.common.def
+
+object PreferenceDef {
+    const val KEY_AD_REMOVAL = "KEY_AD_REMOVAL"
+    const val KEY_HISTORY_DISTANCE = "KEY_HISTORY_DISTANCE"
+    const val KEY_SETTING_LOCATION = "KEY_SETTING_LOCATION"
+    const val KEY_SETTING_THEME = "KEY_SETTING_THEME"
+}
