+++ composeApp/src/commonMain/kotlin/com/yong/taximeter/ui/meter/MeterViewModel.kt
@@ -2,13 +2,13 @@ package com.yong.taximeter.ui.meter
 
 import cafe.adriel.voyager.core.model.ScreenModel
 import cafe.adriel.voyager.core.model.screenModelScope
+import com.yong.taximeter.common.def.PreferenceDef.KEY_SETTING_LOCATION
+import com.yong.taximeter.common.def.PreferenceDef.KEY_SETTING_THEME
 import com.yong.taximeter.common.model.CostInfo
 import com.yong.taximeter.common.model.CostMode
 import com.yong.taximeter.common.util.CostUtil
 import com.yong.taximeter.common.util.LocationUtil
 import com.yong.taximeter.common.util.PreferenceUtil
-import com.yong.taximeter.common.util.PreferenceUtil.KEY_SETTING_LOCATION
-import com.yong.taximeter.common.util.PreferenceUtil.KEY_SETTING_THEME
 import com.yong.taximeter.ui.main.subscreen.setting.model.LocationSetting
 import com.yong.taximeter.ui.main.subscreen.setting.model.ThemeSetting
 import kotlinx.coroutines.Dispatchers
